CLOUDP-120447: Telemetry - Track Events for Interactive Usages
